添加回访图片上传功能，优化请求处理

+++ src/source/webAPI/request.py
@@ -6,22 +6,26 @@ from nicegui import app
 
 def Request(method: str,url: str, data: dict) -> dict:
     headers = {'Content-type': 'application/json','Authorization': app.storage.user.get('TOKEN')}
-    try:
-        if method == 'GET':
-            response = requests.get(url, headers=headers)
-        elif method == 'POST':
-            response = requests.post(url, data=json.dumps(data), headers=headers)
-        elif method == 'PUT':
-            response = requests.put(url, data=json.dumps(data), headers=headers)
-        elif method == 'DELETE':
-            response = requests.delete(url, headers=headers)
+    if data.get("pagesize"):
+        response = requests.get(url, params=data, headers=headers)
+        return response.json()
+    else:
         try:
-            res = response.json()
+            if method == 'GET':
+                response = requests.get(url, headers=headers)
+            elif method == 'POST':
+                response = requests.post(url, data=json.dumps(data), headers=headers)
+            elif method == 'PUT':
+                response = requests.put(url, data=json.dumps(data), headers=headers)
+            elif method == 'DELETE':
+                response = requests.delete(url, headers=headers)
+            try:
+                res = response.json()
+            except:
+                res = {"message":"未知错误","code":0}
+            return res
         except:
-            res = {"message":"未知错误","code":0}
-        return res
-    except:
-        return {"message":"网络错误","code":0}
+            return {"message":"网络错误","code":0}
 
 def GetExcel(url:str, data: dict) -> requests.Response:
     headers = {'Content-type': 'application/json','Authorization': app.storage.user.get('TOKEN')}
